demo/envoy-gateway-benchmark/pyproject.toml
[project]
name = "envoy-gateway-benchmark"
version = "0.1.0"
description = "Statistically rigorous benchmark for Mycelium MCP server"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.1",
    "scipy>=1.12",
    "numpy>=1.26",
]

[tool.hatch.build.targets.wheel]
packages = ["benchlib"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
